7.Advanced Ingestion with llmsherpa/requirements.txt
llama-index
openai
chainlit
pypdf
llmsherpa
pinecone-client
llama-hub
transformers
torch